.history/requirements_20250903095331.txt
Flask==2.3.3
mysql-connector-python==8.1.0
PyMuPDF==1.23.3
python-docx==0.8.11
Werkzeug==2.3.7
gunicorn==21.2.0